Restructure and add filters to features.
- Change psr-4 name from BuiltNorth\\Baseline to WPBaseline.
- Organize code to be more manegable.
- Add filters to most of the core features.
- Update README.md to reflect new features.
