+++ .gitignore
@@ -14,12 +14,6 @@ _srcMaps/
 # Log files
 *.log
 
-#various
-wp-content/uploads/sites/*/gravity_forms
-wp-content/uploads/sites/*/wp-migrate-db
-wp-content/uploads/wpo-plugins-tables-list.json
-
-
 
 # Project Files
 dist
