+++ composer.json
@@ -1,21 +1,25 @@
 {
-    "name": "builtnorth/wp-baseline",
-    "description": "Core functionality that is used on BuiltNorth WordPress sites. Adds security and hardening features and cleans up some default functionality.",
-    "type": "library",
-    "license": "GPL-2.0-or-later",
-    "authors": [
-        {
-            "name": "Built North",
-            "homepage": "https://builtnorth.co"
-        }
-    ],
-    "require": {
-        "php": ">=8.1",
-        "enshrined/svg-sanitize": "^0.20.0"
-    },
-    "autoload": {
-        "psr-4": {
-            "BuiltNorth\\Baseline\\": "inc/"
-        }
-    }
+	"name": "builtnorth/wp-baseline",
+	"description": "Adds baseline functionality to WordPress. For example, security features, admin cleanup, svg sanitization, and more.",
+	"type": "library",
+	"license": "GPL-2.0-or-later",
+	"authors": [
+		{
+			"name": "Built North",
+			"homepage": "https://builtnorth.co"
+		},
+		{
+			"name": "Dan Northern",
+			"homepage": "https://dannorthern.com"
+		}
+	],
+	"require": {
+		"php": ">=8.1",
+		"enshrined/svg-sanitize": "^0.20.0"
+	},
+	"autoload": {
+		"psr-4": {
+			"WPBaseline\\": "inc/"
+		}
+	}
 }
