+++ inc/Cleanup/General.php
@@ -1,0 +1,48 @@
+<?php
+
+/**
+ * ------------------------------------------------------------------
+ * General
+ * ------------------------------------------------------------------
+ *
+ * General cleanup functions
+ *
+ * @package WPBaseline
+ * @since 2.0.0
+ */
+
+namespace WPBaseline\Cleanup;
+
+// Don't load directly.
+defined('ABSPATH') || exit;
+
+class General
+{
+	/**
+	 * Initialize the class.
+	 */
+	public function init()
+	{
+		add_action('init', [$this, 'remove_actions_from_head']);
+	}
+
+
+	/**
+	 * Remove unnecessary code from wp_head
+	 */
+	public function remove_actions_from_head()
+	{
+		remove_action('wp_head', 'rsd_link');
+		remove_action('wp_head', 'wp_generator');
+		remove_action('wp_head', 'feed_links', 2);
+		remove_action('wp_head', 'index_rel_link');
+		remove_action('wp_head', 'wlwmanifest_link');
+		remove_action('wp_head', 'start_post_rel_link');
+		remove_action('wp_head', 'index_rel_link');
+		remove_action('wp_head', 'adjacent_posts_rel_link');
+		remove_action('wp_head', 'feed_links_extra', 3);
+		remove_action('wp_head', 'start_post_rel_link', 10, 0);
+		remove_action('wp_head', 'parent_post_rel_link', 10, 0);
+		remove_action('wp_head', 'adjacent_posts_rel_link_wp_head', 10, 0);
+	}
+}
